Stupid rework of test's teardown method trying to fix one of them
So, the idea is that you can't use the same java class for a collection
and for an embedded document if the class has unique indexes, because
the index is also enforced in the embedding collection and (obviously)
it breaks as soon as you have two similar documents. I'm starting to
think that spring data mongodb has LOTS of limitations.