Create a copy of Language as an inner class of User
After the unfortunate fiasco of trying to reuse a class with indexes,
only two options came out of my mind: "denormalize" the model so the
classes are equal but different, and revert to the original plan of
using a DBRef. While I'm more comfortable with the more "relational"
approach of the latter, I went with the former because fuck the system.