Add a simple event-driven Post lifecycle management feature
The lifecycle of a Post is something that a few subsystems are usually
interested in. A canonical example would be cache invalidation.

There's a few ways to implement this, but spring's built-in event
processing mechanism fits well considering the increasing amount of
functionality that will plug in, e.g. tag counts, archive entry counts,
search index, different caching layers...